requirements.txt
django-cors-headers==0.11
Django==1.5.1
argparse==1.2.1
django-extensions==1.1.1
django-stdimage==0.2.2
pycrypto==2.6
wsgiref==0.1.2
